Introduce factory method for provisioning instances
