+++ src/main/scala/mesosphere/marathon/core/instance/Instance.scala
@@ -1,11 +1,13 @@
 package mesosphere.marathon
 package core.instance
 
+import java.time.Clock
 import java.util.{Base64, UUID}
 
 import com.fasterxml.uuid.{EthernetAddress, Generators}
 import mesosphere.marathon.core.condition.Condition
 import mesosphere.marathon.core.instance.Instance.{AgentInfo, InstanceState}
+import mesosphere.marathon.core.pod.PodDefinition
 import mesosphere.marathon.core.task.Task
 import mesosphere.marathon.state._
 import mesosphere.marathon.tasks.OfferUtil
@@ -13,7 +15,7 @@ import mesosphere.marathon.stream.Implicits._
 import mesosphere.mesos.Placed
 import mesosphere.marathon.raml.Raml
 import org.apache._
-import org.apache.mesos.Protos.Attribute
+import org.apache.mesos.Protos.{Attribute, NetworkInfo}
 import play.api.libs.json._
 import play.api.libs.json.Reads._
 import play.api.libs.functional.syntax._
@@ -116,6 +118,36 @@ object Instance {
      * @return An instance in the scheduled state.
      */
     def apply(runSpec: RunSpec): Instance = Scheduled(runSpec, Id.forRunSpec(runSpec.id))
+  }
+
+  object Provisioned {
+    /**
+      * Factory method for creating provisioned instance from Scheduled instance for apps
+      * @param scheduledInstance instance in a Scheduled state
+      * @return new instance in a provisioned state
+      */
+    def apply(scheduledInstance: Instance, agentInfo: AgentInfo, networkInfo: core.task.state.NetworkInfo, app: AppDefinition)(implicit clock: Clock): Instance = {
+      require(scheduledInstance.isScheduled, "Scheduled instance is required to create provisioned instance.")
+
+      val task = Task(
+        taskId = Task.Id.forInstanceId(scheduledInstance.instanceId, None),
+        runSpecVersion = app.version,
+        status = Task.Status(
+          stagedAt = clock.now(),
+          condition = Condition.Created,
+          networkInfo = networkInfo
+        )
+      )
+
+      val tasksMap = Map(task.taskId -> task)
+
+      scheduledInstance.copy(
+        agentInfo = Some(agentInfo),
+        state = Instance.InstanceState(Condition.Provisioned, clock.now(), None, None),
+        tasksMap = tasksMap,
+        runSpecVersion = app.version // Note: The app version might change.
+      )
+    }
   }
 
   /**
